Moved data classes into java-data
