+++ .github/ISSUE_TEMPLATE/BUG_REPORT.md
@@ -1,7 +1,5 @@
 ---
-name: Bug Report
-about: Create a bug report to help us improve Cryptic.
-labels: bug
+name: Bug Report about: Create a bug report to help us improve Cryptic. labels: bug
 
 ---
