+++ admin-panel/src/main/java/net/cryptic_game/backend/admin/AdminPanelBootstrap.java
@@ -1,9 +1,9 @@
 package net.cryptic_game.backend.admin;
 
 import lombok.extern.slf4j.Slf4j;
-import net.cryptic_game.backend.admin.authentication.Group;
-import net.cryptic_game.backend.admin.data.sql.entities.user.AdminUser;
-import net.cryptic_game.backend.admin.data.sql.repositories.user.AdminUserRepository;
+import net.cryptic_game.backend.data.Group;
+import net.cryptic_game.backend.data.sql.entities.user.AdminUser;
+import net.cryptic_game.backend.data.sql.repositories.user.AdminUserRepository;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
@@ -11,7 +11,6 @@ import java.util.Set;
 
 @Slf4j
 @Configuration
-@EnableJpaRepositories(basePackages = "net.cryptic_game.backend.admin.data.sql")
 public class AdminPanelBootstrap {
 
     public AdminPanelBootstrap(final AdminPanelConfig config,
