+++ admin-panel/src/main/java/net/cryptic_game/backend/admin/endpoints/authentication/AdminGroupsEndpoints.java
@@ -2,13 +2,13 @@ package net.cryptic_game.backend.admin.endpoints.authentication;
 
 import io.netty.handler.codec.http.HttpResponseStatus;
 import net.cryptic_game.backend.admin.authentication.AdminPanelAuthenticator;
-import net.cryptic_game.backend.admin.authentication.Group;
-import net.cryptic_game.backend.admin.authentication.Permission;
 import net.cryptic_game.backend.base.api.annotations.ApiEndpoint;
 import net.cryptic_game.backend.base.api.annotations.ApiEndpointCollection;
 import net.cryptic_game.backend.base.api.data.ApiResponse;
 import net.cryptic_game.backend.base.api.data.ApiType;
 import net.cryptic_game.backend.base.json.JsonBuilder;
+import net.cryptic_game.backend.data.Group;
+import net.cryptic_game.backend.data.Permission;
 
 import java.util.Arrays;
 import java.util.stream.Collectors;
