+++ admin-panel/src/main/java/net/cryptic_game/backend/admin/endpoints/authentication/AdminUserEndpoints.java
@@ -4,10 +4,6 @@ import com.google.gson.JsonObject;
 import io.netty.handler.codec.http.HttpResponseStatus;
 import lombok.RequiredArgsConstructor;
 import net.cryptic_game.backend.admin.authentication.AdminPanelAuthenticator;
-import net.cryptic_game.backend.admin.authentication.Group;
-import net.cryptic_game.backend.admin.authentication.Permission;
-import net.cryptic_game.backend.admin.data.sql.entities.user.AdminUser;
-import net.cryptic_game.backend.admin.data.sql.repositories.user.AdminUserRepository;
 import net.cryptic_game.backend.base.api.annotations.ApiEndpoint;
 import net.cryptic_game.backend.base.api.annotations.ApiEndpointCollection;
 import net.cryptic_game.backend.base.api.annotations.ApiParameter;
@@ -16,6 +12,10 @@ import net.cryptic_game.backend.base.api.data.ApiResponse;
 import net.cryptic_game.backend.base.api.data.ApiType;
 import net.cryptic_game.backend.base.api.handler.rest.RestApiRequest;
 import net.cryptic_game.backend.base.json.JsonUtils;
+import net.cryptic_game.backend.data.Group;
+import net.cryptic_game.backend.data.Permission;
+import net.cryptic_game.backend.data.sql.entities.user.AdminUser;
+import net.cryptic_game.backend.data.sql.repositories.user.AdminUserRepository;
 
 import java.util.Objects;
 import java.util.Optional;
