+++ admin-panel/src/main/java/net/cryptic_game/backend/admin/endpoints/authentication/AuthenticationTokenEndpoints.java
@@ -3,8 +3,8 @@ package net.cryptic_game.backend.admin.endpoints.authentication;
 import com.google.gson.JsonObject;
 import io.netty.handler.codec.http.HttpResponseStatus;
 import lombok.RequiredArgsConstructor;
-import net.cryptic_game.backend.admin.data.sql.entities.user.AdminUser;
-import net.cryptic_game.backend.admin.data.sql.repositories.user.AdminUserRepository;
+import net.cryptic_game.backend.data.sql.entities.user.AdminUser;
+import net.cryptic_game.backend.data.sql.repositories.user.AdminUserRepository;
 import net.cryptic_game.backend.base.api.annotations.ApiEndpoint;
 import net.cryptic_game.backend.base.api.annotations.ApiEndpointCollection;
 import net.cryptic_game.backend.base.api.annotations.ApiParameter;
