+++ admin-panel/src/main/java/net/cryptic_game/backend/admin/endpoints/authentication/OAuthEndpoints.java
@@ -9,8 +9,6 @@ import io.netty.handler.codec.http.HttpResponseStatus;
 import lombok.RequiredArgsConstructor;
 import net.cryptic_game.backend.admin.authentication.AuthenticationErrorException;
 import net.cryptic_game.backend.admin.authentication.OAuthConfig;
-import net.cryptic_game.backend.admin.data.sql.entities.user.AdminUser;
-import net.cryptic_game.backend.admin.data.sql.repositories.user.AdminUserRepository;
 import net.cryptic_game.backend.base.api.annotations.ApiEndpoint;
 import net.cryptic_game.backend.base.api.annotations.ApiEndpointCollection;
 import net.cryptic_game.backend.base.api.annotations.ApiParameter;
@@ -19,6 +17,8 @@ import net.cryptic_game.backend.base.api.data.ApiType;
 import net.cryptic_game.backend.base.json.JsonBuilder;
 import net.cryptic_game.backend.base.json.JsonUtils;
 import net.cryptic_game.backend.base.utils.SecurityUtils;
+import net.cryptic_game.backend.data.sql.entities.user.AdminUser;
+import net.cryptic_game.backend.data.sql.repositories.user.AdminUserRepository;
 import reactor.core.publisher.Mono;
 import reactor.netty.http.client.HttpClient;
