+++ admin-panel/src/main/java/net/cryptic_game/backend/admin/endpoints/website/team/TeamMemberEndpoints.java
@@ -3,16 +3,16 @@ package net.cryptic_game.backend.admin.endpoints.website.team;
 import io.netty.handler.codec.http.HttpResponseStatus;
 import lombok.RequiredArgsConstructor;
 import net.cryptic_game.backend.admin.authentication.AdminPanelAuthenticator;
-import net.cryptic_game.backend.admin.authentication.Permission;
-import net.cryptic_game.backend.admin.data.sql.entities.website.team.TeamDepartment;
-import net.cryptic_game.backend.admin.data.sql.entities.website.team.TeamMember;
-import net.cryptic_game.backend.admin.data.sql.repositories.website.team.TeamDepartmentRepository;
-import net.cryptic_game.backend.admin.data.sql.repositories.website.team.TeamMemberRepository;
 import net.cryptic_game.backend.base.api.annotations.ApiEndpoint;
 import net.cryptic_game.backend.base.api.annotations.ApiEndpointCollection;
 import net.cryptic_game.backend.base.api.annotations.ApiParameter;
 import net.cryptic_game.backend.base.api.data.ApiResponse;
 import net.cryptic_game.backend.base.api.data.ApiType;
+import net.cryptic_game.backend.data.Permission;
+import net.cryptic_game.backend.data.sql.entities.website.team.TeamDepartment;
+import net.cryptic_game.backend.data.sql.entities.website.team.TeamMember;
+import net.cryptic_game.backend.data.sql.repositories.website.team.TeamDepartmentRepository;
+import net.cryptic_game.backend.data.sql.repositories.website.team.TeamMemberRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
