+++ java-base/src/main/java/net/cryptic_game/backend/base/sql/models/TableModelAutoId.java
@@ -3,7 +3,6 @@ package net.cryptic_game.backend.base.sql.models;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.hibernate.annotations.GenericGenerator;
 import org.hibernate.annotations.Type;
 
 import javax.persistence.Column;
@@ -21,9 +20,8 @@ public abstract class TableModelAutoId extends TableModel {
 
     @Id
     @Type(type = "uuid-char")
-    @GeneratedValue(generator = "UUID")
+    @GeneratedValue
     @Column(name = "id", updatable = false, nullable = false)
-    @GenericGenerator(name = "UUID", strategy = "org.hibernate.id.UUIDGenerator")
     private UUID id;
 
     /**
