+++ java-data/src/main/java/net/cryptic_game/backend/data/sql/entities/website/team/TeamMember.java
@@ -1,4 +1,4 @@
-package net.cryptic_game.backend.admin.data.sql.entities.website.team;
+package net.cryptic_game.backend.data.sql.entities.website.team;
 
 import com.google.gson.JsonElement;
 import lombok.AllArgsConstructor;
