+++ java-data/src/main/java/net/cryptic_game/backend/data/sql/repositories/user/AdminUserRepository.java
@@ -1,6 +1,6 @@
-package net.cryptic_game.backend.admin.data.sql.repositories.user;
+package net.cryptic_game.backend.data.sql.repositories.user;
 
-import net.cryptic_game.backend.admin.data.sql.entities.user.AdminUser;
+import net.cryptic_game.backend.data.sql.entities.user.AdminUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AdminUserRepository extends JpaRepository<AdminUser, Long> {
